refactor: move dtos folder into routers
